+++ packages/bluesky/CHANGELOG.md
@@ -4,9 +4,12 @@
 
 - Fixed standard prefix from `destroy` to `delete`. ([#81](https://github.com/myConsciousness/atproto.dart/issues/81))
 - Fixed standard prefix from `lookup` to `get`. ([#84](https://github.com/myConsciousness/atproto.dart/issues/84))
-- Add methods in `FeedsService`. ([#33](https://github.com/myConsciousness/atproto.dart/issues/33))
+- Added methods in `FeedsService`. ([#33](https://github.com/myConsciousness/atproto.dart/issues/33))
   - `createLike`
   - `deleteLike`
+- Added `GraphsService` and methods to follow users. ([#29](https://github.com/myConsciousness/atproto.dart/issues/29))
+  - `createFollow`
+  - `deleteFollow`
 
 ## v0.2.0